requirements.txt
gitpython
emoji
colorama
click
